Move home.tsx to components, remove configurations
