+++ app/page.tsx
@@ -1,4 +1,4 @@
-import Home from "./home/page";
+import Home from "app/components/home";
 
 const Page = () => {
   return (
